Succeded

+++ common/network/include/message.h
@@ -35,7 +35,7 @@ struct message {
     }
 
     friend std::ostream &operator<<(std::ostream &out, const message &msg) {  // in case of debugging
-        out << "Message (m_id=" << msg.header.id << "): size=" << msg.size() << "\n";
+        out << "Message (m_id=" << (int)msg.header.id << "): size=" << msg.size() << "\n";
         return out;
     }
